Change: Use TypeScript for ErrorPanel component
